Java 5 type checking and compiler warnings
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3710 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/compatibility/ValueInteger.java
@@ -65,7 +65,7 @@ public class ValueInteger implements ValueInterface, Cloneable
 
 	public double getNumber()
 	{
-		return (double)this.number;
+		return this.number;
 	}
 
 	public Date getDate()
